chore(items): simplified data handling a lot
Initially the data validity was meant to be checked when constructing,
but some of the deciding factor can only be checked when encoding.

Now the default is to check at encoding time, but some check can
still be done on creation time, e.g. for Domain most of the checks.