refactor(codec/ecw): buf u8 -> AsciiChar
Before a u8 tmp. buffer was used with unsafe conversion to AsciiChar,
when it was written to the output, not the tmp. buffer is Ascii and
the unsafe conversion is done when writing to the bufer, through
at some places the data was already a AsciiChar not needing any
conversion