chore(resource): Rewrote Resource to abstract over the inner state
Before a resource could have been in one of multiple states, a
specification where/how to load it, a future loading it, a
a buffer, transfer encoded buffer, etc. and the code using
the type had to handle this.

Now it contains this state internally only exposing
a way to advance it ( `as_future` ) and retrive a
reference to a Transfer Encoded buffer if the
state has advanced to that point ( `get_if_encoded` ).

Also this inner state is wrapped by a `Arc<Lock<>>` making
a `Resource` cheap to clone AND making all work like loading
and encoding it is only done once for all clones.