chore(encoder): integrated new encoder until compliation was possible
But 30 (of 216) test _fail now_, mainly because of the change in the
trace/debug tokens in the `ec_test` macro.

There is a number of other task which have to be done to complete
the refactoring including some renaming and making the drop-panic
check of the HeaderEncoderHandle #[cfg(test)] only, providing a
`with_encoding_handle` function on `Encoder` which handles the
Handle end of life parts like emiting a error if there are
unfinished parts etc.