chore(mime_sniff): initial implementation
Works for many cases but fails hard on some due to drawbacks of the
`tree_magic` library. Mainly there is a problem with files which
can be more than one type (kind of a common occurence, e.g. a pdf
can be `test/plain; charset=utf8` and therefore also be `text/x-tex`).
It doesn't use the weights of libmagic at all....