fix(internals): fixed latend safety bug
The function for efficiently inserting a slice of bytes into
a `Vec<u8>` did miss a bounds check and was, to make thinks
worse optimized using unsafe code. Still the bug was only
latent as all parts using the function allways had in bounds
indexes. (There was only one, which was in a grow only buffer
and used a index which is always set from the current buffer
len, which as noted can only grow).