fix(header/phrase): Fixed bug which did eat spaces between some words.
Closes #14

The new algorithm either represents a the phrase through a single
quoted string or through a sequence of encoded words. With this
it's less "nice" then the previous algorithms but it also means
we don't have to check for text looking like encoded words, nor
do we have to do a look-ahead/look-behind to prevent ws from being
eaten, nor do we have to add some code to decide when some "free"
chars should be moved into the encoded-word or when they should
be represented seperateley.